make reusable login form Relates #84
