+++ client/src/pages/signin/index.jsx
@@ -1,11 +1,9 @@
-import React, { useState } from 'react';
+import React from 'react';
+
 import {
-  Form,
   Button,
-  Input,
-  message,
   Typography,
-  Anchor,
+  message,
   Space,
 } from 'antd';
 import { useDispatch } from 'react-redux';
@@ -15,32 +13,14 @@ import axios from 'axios';
 
 import { sign } from '../../redux/feature/user/userSlice';
 import '../signup/style.less';
+import SignForm from '../../components/common/SignForm';
 
-const { Link } = Anchor;
-const { Password } = Input;
-const { Item } = Form;
-const { Title, Text } = Typography;
+const { Title } = Typography;
 
 function Signin() {
   const dispatch = useDispatch();
-  const [userInfo, setUserInfo] = useState({
-    email: '', password: '',
-  });
 
-  const passwordValidation = () => ({
-    validator(_, value) {
-      if (value.length >= 6) {
-        return Promise.resolve();
-      }
-      return Promise.reject(new Error('password should have at least 6 character '));
-    },
-  });
-
-  const handleChange = ({ target: { name, value } }) => {
-    setUserInfo({ ...userInfo, [name]: value });
-  };
-
-  const login = async () => {
+  const login = async (userInfo) => {
     try {
       const { data: { data } } = await axios.post('/api/login', userInfo);
       dispatch(sign(data));
@@ -73,60 +53,7 @@ function Signin() {
           >
             LOGIN
           </Title>
-          <Form
-            className="Form-sign-up"
-            name="register"
-            onFinish={login}
-          >
-            <Item
-              name="email"
-              rules={[
-                {
-                  type: 'email',
-                  message: 'The input is not valid E-mail!',
-                },
-                {
-                  required: true,
-                  message: 'Please input your E-mail!',
-                },
-              ]}
-            >
-              <Input
-                name="email"
-                placeholder="Email"
-                onChange={(e) => handleChange(e)}
-              />
-            </Item>
-
-            <Item
-              name="password"
-              rules={[
-                {
-                  required: true,
-                  message: 'Please input your password!',
-                },
-                () => passwordValidation(),
-              ]}
-            >
-              <Password
-                placeholder="Password"
-                name="password"
-                onChange={(e) => handleChange(e)}
-              />
-            </Item>
-            <Button className="sign-up-btn" type="primary" htmlType="submit">
-              Sign In
-            </Button>
-            <div className="register_option">
-              <Text>Don`t have an account ?</Text>
-              <Anchor affix={false}>
-                <Link
-                  href="/signup"
-                  title="Sign up"
-                />
-              </Anchor>
-            </div>
-          </Form>
+          <SignForm getUserInfo={login} />
           <Button type="primary" icon={<GoogleOutlined />}>
             Sign in with Google
           </Button>
